add Simple Sprite Batcher
add Simple Sprite Batcher as well an example program along with its texture.

to use the batcher, simply pass all the sprites to it and then draw the batcher instead.
you can pass a vector of sprites (useful if you store your sprites in a vector) or a vector of pointers to sprites (useful for everything else). you can create a vector of pointers easily and it shouldn't even affect performance if you do it every frame, if required. you could, of course, use this to order them as well.